+++ release_notes.md
@@ -1,5 +1,6 @@
-## v4.1
-- ?
+## v4.1 (unreleased)
+- `app`, `test`, and `xctestrun-file` now support `~`, environment variables, and globs (`*`, `**`) when resolving paths. #386
+- Update `flank android run` to support `--app`, `--test`, `--test-targets`, `--use-orchestrator` and `--no-use-orchestrator`.
 
 ## v4.0.0